+++ src/supervision/key_points/core.py
@@ -21,22 +21,15 @@ from supervision.validators import _validate_keypoints_fields
 logger = logging.getLogger(__name__)
 
 Index1D = Union[
-    int,
-    slice,
-    list[int],
-    list[bool],
-    npt.NDArray[np.int_],
-    npt.NDArray[np.bool_],
+    int, slice, list[int], list[bool], npt.NDArray[np.int_], npt.NDArray[np.bool_]
 ]
 Index2D = tuple[Index1D, Index1D]
-_RowIndexInput = Union[
-    int,
-    np.integer[Any],
+_RowIndexInput = Union[int, np.integer[Any], npt.NDArray[np.generic], list[Any], slice]
+_NormalizedRowIndex = Union[
     npt.NDArray[np.generic],
     list[Any],
     slice,
 ]
-_NormalizedRowIndex = Union[npt.NDArray[np.generic], list[Any], slice]
 
 
 def _optional_array_equal(